Fix String Compatibility Issues in BleMouse.cpp Library
This commit addresses the compatibility issues between std::string and Arduino String types in the BleMouse library. Modifications have been made to `BleMouse.cpp` to ensure that std::string is properly converted to Arduino String where necessary. These changes ensure that the library functions correctly with Arduino IDE Version 2.2.1 and the developer version of the Arduino-ESP32 library (Arduino Alpha 2 Release v3.0.0 based on ESP-IDF v5.1). The library has been tested and verified to work as expected with this setup.